build: rename files for linux build
Linux is case-sensative, so CMake threw an error during generation
because it couldn't find the right files.
I've decided to rename the files rather than change their names in CMake
for consistency in naming all the files.
