+++ src/abstractsolver.cpp
